+++ db/migrate/20141121190307_create_questions.rb
@@ -2,8 +2,8 @@ class CreateQuestions < ActiveRecord::Migration
   def change
     create_table :questions do |t|
       t.belongs_to :notation
-      t.integer :number
-      t.integer :mark
+      t.double :number
+      t.double :mark
       t.integer :points
       t.text :comment
